Get productivity rpoets from api

+++ app/src/main/java/com/example/chaquitaclla_appmovil_android/statistics/beans/ProductivityReport.kt
@@ -3,7 +3,7 @@ package com.example.chaquitaclla_appmovil_android.statistics.beans
 class ProductivityReport(
     var id: Int,
     var name: String,
-    var production: Int,
-    var area: Int
+    var production: Float,
+    var area: Float
 ) {
 }
